IN-196 refactor: add index
