+++ src/features/createPost/model/adding/index.ts
@@ -1,3 +1,4 @@
 export * from './addPhotoSchema'
 export * from './addPhotosCheck'
 export * from './useAddPhotoModal'
+export * from './useNewPostDialog'
