+++ src/features/createPost/model/cropping/index.ts
@@ -1,2 +1,3 @@
 export * from './useAddPhotoControllerPopover'
 export * from './usePhotoCrop'
+export * from './useScaleControllerPopover'
