+++ src/features/createPost/model/index.ts
@@ -3,3 +3,4 @@ export * from './adding'
 export * from './cropping'
 export * from './publication'
 export * from './types'
+export * from './utils'
