+++ src/features/createPost/model/publication/index.ts
@@ -1,1 +1,2 @@
 export * from './useCityAutocomplete'
+export * from './usePublicationModal'
