+++ src/features/createPost/ui/adding/index.ts
@@ -1,0 +1,2 @@
+export * from './AddPhotoModal'
+export * from './NewPostDialog'
