+++ src/features/createPost/ui/components/index.ts
@@ -1,4 +1,4 @@
-export * from './ModalHeader'
-export * from './SwiperCover'
+export * from './modalHeader'
 export * from './photoControllerButton'
 export * from './slider'
+export * from './swiperCover'
