+++ src/features/createPost/ui/components/modalHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './ModalHeader'
