+++ src/features/createPost/ui/components/swiperCover/index.ts
@@ -1,0 +1,1 @@
+export * from './SwiperCover'
