+++ src/features/createPost/ui/cropping/ControllersPanel/ScaleControllerPopover.tsx
@@ -12,8 +12,8 @@ import {
   PhotoControllerButton,
   Slider,
   addPostStore,
+  useScaleControllerPopover,
 } from '@/features/createPost'
-import { useScaleControllerPopover } from '@/features/createPost/model/cropping/useScaleControllerPopover'
 import { observer } from 'mobx-react-lite'
 
 type Props = {
